Emit block index with feed 'update' event
This event is not documented, but I assume from the context that it will emit whenever the feed is changed, ie. data is appended. I intend of using this in [`hypercore-sparse-index`](https://github.com/emilbayes/hypercore-sparse-index) to keep the index up to date whenever data is added to the feed.

If this is not going to be part of the public API, or if I misunderstood its purpose, just close :smile: